store user info in session

+++ HabrLessonClassLibrary/Domain/User.cs
@@ -2,12 +2,14 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using System.Numerics;
+
 
 namespace HabrLessonClassLibrary.Domain
 {
     public class User
     {
-        public int Id { get; set; }
+        public BigInteger Id { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string LinkToAvatar { get; set; }
